+++ inc/Core/Converter.php
@@ -84,6 +84,9 @@ class Converter {
 		$this->set_image_source();
 		$this->set_image_destination();
 
+		// Set response.
+		$webp = $this->rel_dest;
+
 		// Bail out, if source is empty.
 		if ( ! file_exists( $this->abs_source ) ) {
 			return new \WP_Error(
@@ -111,7 +114,7 @@ class Converter {
 
 		// Bail out, if dest. image exists.
 		if ( file_exists( $this->abs_dest ) ) {
-			return $this->rel_dest;
+			return $webp;
 		}
 
 		// Convert to WebP image.
@@ -124,8 +127,7 @@ class Converter {
 				$e->getMessage()
 			);
 			error_log( $error_msg );
-
-			return new \WP_Error( 'webp-img-error', $error_msg );
+			$webp = new \WP_Error( 'webp-img-error', $error_msg );
 		}
 
 		/**
@@ -133,15 +135,16 @@ class Converter {
 		 *
 		 * @since 1.0.1
 		 * @since 1.1.0 Moved to Converter.
+		 * @since 1.1.1 Rename hook to use `icfw` prefix.
 		 *
 		 * @param string|\WP_Error $webp          WebP Image URL or WP Error.
 		 * @param int              $attachment_id Image ID.
 		 *
 		 * @return void
 		 */
-		do_action( 'webp_img_convert', $webp = $this->rel_dest, $attachment_id = $this->service->source['id'] );
+		do_action( 'icfw_convert', $webp, $attachment_id = $this->service->source['id'] );
 
-		return $this->rel_dest;
+		return $webp;
 	}
 
 	/**
@@ -191,7 +194,7 @@ class Converter {
 	 * @return mixed[]
 	 */
 	protected function get_options(): array {
-		$settings = get_option( 'webp_img_converter', [] );
+		$settings = get_option( 'icfw', [] );
 
 		// Make sure this array key is integer.
 		$settings['quality'] = (int) ( $settings['quality'] ?? 0 );
@@ -210,10 +213,11 @@ class Converter {
 		 *
 		 * @since 1.0.0
 		 * @since 1.1.0 Moved to Converter.
+		 * @since 1.1.1 Rename hook to use `icfw` prefix.
 		 *
 		 * @param mixed[] $options Conversion options.
 		 * @return mixed[]
 		 */
-		return (array) apply_filters( 'webp_img_options', $options );
+		return (array) apply_filters( 'icfw_options', $options );
 	}
 }