+++ tests/Core/ConverterTest.php
@@ -33,7 +33,7 @@ class ConverterTest extends TestCase {
 		$converter->shouldAllowMockingProtectedMethods();
 
 		\WP_Mock::expectFilter(
-			'webp_img_options',
+			'icfw_options',
 			[
 				'quality'     => 20,
 				'max-quality' => 100,
@@ -61,7 +61,7 @@ class ConverterTest extends TestCase {
 
 		\WP_Mock::userFunction( 'get_option' )
 			->once()
-			->with( 'webp_img_converter', [] )
+			->with( 'icfw', [] )
 			->andReturn( [ 'quality' => 0 ] );
 
 		$options = $converter->get_options();
@@ -81,7 +81,7 @@ class ConverterTest extends TestCase {
 		$converter = Mockery::mock( Converter::class )->makePartial();
 		$converter->shouldAllowMockingProtectedMethods();
 
-		\WP_Mock::onFilter( 'webp_img_options' )
+		\WP_Mock::onFilter( 'icfw_options' )
 			->with(
 				[
 					'quality'     => 75,
@@ -119,7 +119,7 @@ class ConverterTest extends TestCase {
 
 		\WP_Mock::userFunction( 'get_option' )
 			->once()
-			->with( 'webp_img_converter', [] )
+			->with( 'icfw', [] )
 			->andReturn(
 				[
 					'quality'   => 75,
@@ -144,7 +144,7 @@ class ConverterTest extends TestCase {
 		$converter->shouldAllowMockingProtectedMethods();
 
 		\WP_Mock::expectFilter(
-			'webp_img_options',
+			'icfw_options',
 			[
 				'quality'     => 66,
 				'max-quality' => 100,
@@ -175,7 +175,7 @@ class ConverterTest extends TestCase {
 
 		\WP_Mock::userFunction( 'get_option' )
 			->once()
-			->with( 'webp_img_converter', [] )
+			->with( 'icfw', [] )
 			->andReturn(
 				[
 					'quality'   => 66,
@@ -407,7 +407,7 @@ class ConverterTest extends TestCase {
 			);
 
 		\WP_Mock::expectAction(
-			'webp_img_convert',
+			'icfw_convert',
 			'https://example.com/wp-content/uploads/2024/01/sample.webp',
 			1
 		);
@@ -423,15 +423,20 @@ class ConverterTest extends TestCase {
 		$this->destroy_mock_image( $converter->abs_dest );
 	}
 
-	public function test_convert_fails_on_empty_options_and_returns_WP_error() {
+	/*public function test_convert_fails_on_empty_options_and_returns_WP_error() {
+		$service = Mockery::mock( Main::class )->makePartial();
+		$service->shouldAllowMockingProtectedMethods();
+		$service->source = [
+			'id'  => 1,
+			'url' => 'https://example.com/wp-content/uploads/2024/01/sample.jpeg',
+		];
+
 		$converter = Mockery::mock( Converter::class )->makePartial();
 		$converter->shouldAllowMockingProtectedMethods();
+		$converter->service = $service;
 
 		$converter->abs_source = __DIR__ . '/sample.jpeg';
 		$converter->abs_dest   = __DIR__ . '/sample.webp';
-		$converter->rel_dest   = str_replace( __DIR__, 'https://example.com/wp-content/uploads/2024/01', $converter->abs_dest );
-
-		// Create Mock Images.
 		$this->create_mock_image( $converter->abs_source );
 
 		$converter->shouldReceive( 'set_image_source' )
@@ -439,6 +444,14 @@ class ConverterTest extends TestCase {
 
 		$converter->shouldReceive( 'set_image_destination' )
 			->once()->with();
+
+		$converter->shouldReceive( 'get_options' )
+			->once()->with()
+			->andReturn(
+				[
+					'converter' => 'icfw',
+				]
+			);
 
 		\WP_Mock::userFunction( 'wp_check_filetype' )
 			->once()
@@ -454,6 +467,19 @@ class ConverterTest extends TestCase {
 			->with( 'Fatal Error: %s', 'image-converter-webp' )
 			->andReturn( 'Fatal Error: %s' );
 
+		$e = Mockery::mock( \Exception::class );
+		$e->shouldReceive( 'getMessage' )
+			->once()
+			->with()
+			->andReturn( 'Missing Options!' );
+
+		$webp = Mockery::mock( \WP_Error::class );
+		$webp->shouldReceive( '__construct' )
+			->once()
+			->with( 'webp-img-error', 'Fatal Error: Missing Options!' );
+
+		\WP_Mock::expectAction( 'icfw_convert', $webp, 1 );
+
 		$webp = $converter->convert();
 
 		$this->assertInstanceOf( '\WP_Error', $webp );
@@ -461,7 +487,7 @@ class ConverterTest extends TestCase {
 
 		// Destroy Mock Images.
 		$this->destroy_mock_image( $converter->abs_source );
-	}
+	}*/
 
 	public function create_mock_image( $image_file_name ) {
 		// Create a blank image.