+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.1
+* Ensure WP_Error is passed and returned to Hook.
+* Rename hooks across codebase to use `icfw` prefix.
+* Implement Kernel interface.
+* Fix bugs & failing tests.
+* Update README notes.
+
 ## 1.1.0
 * Major code base refactor.
 * Add more **Settings** options to **Settings** page.
@@ -29,7 +36,7 @@
 * Add more Unit tests & Code coverage.
 
 ## 1.0.2
-* Add `webp_img_delete` and `webp_img_metadata_delete` hooks.
+* Add `icfw_delete` and `icfw_metadata_delete` hooks.
 * Add Settings page for plugin options.
 * Add WebP field on WP attachment modal.
 * Add new class methods.
@@ -38,13 +45,13 @@
 * Update README notes.
 
 ## 1.0.1
-* Refactor hook `webp_img_convert` to placement within convert public method.
+* Refactor hook `icfw_convert` to placement within convert public method.
 * Fix bugs within `Plugin` class methods.
 * Add more Unit tests & Code coverage.
 * Update README notes.
 
 ## 1.0.0 (Initial Release)
 * WebP image conversion for any type of image.
-* Custom Hooks - `webp_img_options`, `webp_img_convert`, `webp_img_attachment_html`, `webp_img_thumbnail_html`.
+* Custom Hooks - `icfw_options`, `icfw_convert`, `icfw_attachment_html`, `icfw_thumbnail_html`.
 * Unit Tests coverage.
 * Tested up to WP 6.5.3.